Handle starting the camera in landscape on tablets
This commit also ensures that the CameraPreviewController does not return the Camera to the CameraPreviewTextureView. Instead only a future of a PreviewSize is returned, and this allows the TextureView to resize itself properly whenever a new camera is opened with a potentially different PreviewSize.

All rotation of the CameraFragment is now handled inside of the Fragment itself. If the Camera feed is showing, the orientation is locked into portrait, else it is freed up again to let the device decide.